pyproject.toml
[tool.poetry]
name = "cruft_helloworld"
version = "0.2.0"
description = "Cruft Python Hello-World"
authors = [
    "lionel atty <yoyonel@hotmail.com>",
]

license = "BSD-3-Clause"
readme = "README.md"

documentation = "https://github.com/yoyonel/cruft_helloworld/blob/master/README.md"
homepage = "https://github.com/yoyonel/cruft_helloworld"
repository = "https://github.com/yoyonel/cruft_helloworld"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

include = ["cruft_helloworld/py.typed"]
packages = [
    { include = "cruft_helloworld" },
]

# https://python-poetry.org/docs/pyproject/#scripts
[tool.poetry.scripts]
# https://click.palletsprojects.com/en/5.x/setuptools/#scripts-in-packages
helloworld = "cruft_helloworld.app:cli"

[tool.poetry.dependencies]
python = ">=3.7, <4.0"
colorlog = "^4.6.2"
click = "^7.1.2"
click-default-group = "^1.2.2"
Cython = "^0.29.21"
python-geoip-geolite2 = "^2015.0303"
python-geoip-python3 = "^1.3"
rich = "^9.4.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.3"
ipython = "^7.19.0"
isort = "^5.5.2"
mypy = "^0.782"
pre-commit = "^2.7.1"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"

tox = "^3.20.0"
tox-gh-actions = "^1.3.0"
pytest-click = "^1.0.2"
pytest-parametrization = "^2019.1.4"
tox-pyenv = "^1.1.0"
lxml = "^4.6.2"

[tool.black]
line-length = 88
skip-string-normalization = true
target_version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = ["using_geoip"]
addopts = "--doctest-modules"
